+++ HrApp.Domain/Entities/Offer.cs
@@ -6,7 +6,7 @@ public class Offer
     public Guid TeamId { get; set; } 
     public virtual Team Team { get; set; } = default!; 
     public string PositionName { get; set; } = default!;
-    public float Salary { get; set; }
+    public double Salary { get; set; }
     public string Description { get; set; } = default!;
     public DateOnly AddDate { get; set; }
     public List<JobApplication> JobApplications { get; set; } = default!;
